% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Get example dataset (AOI + NDVI time series)}
\usage{
get_example_data()
}
\value{
A list with:
\describe{
\item{aoi}{An sf polygon object}
\item{files}{Character vector with NDVI raster file paths}
\item{dates}{Date vector corresponding to the rasters}
}
}
\description{
Loads the synthetic example data shipped with the package.
}
