% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_local.R
\name{get_index_local}
\alias{get_index_local}
\title{Read NDVI/NBR time series from local files and clip to AOI}
\usage{
get_index_local(aoi, files, dates)
}
\arguments{
\item{aoi}{sf polygon (Area of Interest)}

\item{files}{character vector of raster file paths (one per date)}

\item{dates}{Date vector (same length as files)}
}
\value{
A terra SpatRaster with one layer per date (names = dates)
}
\description{
Read NDVI/NBR time series from local files and clip to AOI
}
